[8114] Fixed uint16 to uint32 for spell id in SkillDiscoveryEntry
(cherry picked from commit 0f5471ea2a07fd78f7b52dc9d964eb2ecd6b0725)

+++ src/game/SkillDiscovery.cpp
@@ -35,7 +35,7 @@ struct SkillDiscoveryEntry
     SkillDiscoveryEntry()
         : spellId(0), chance(0) {}
 
-    SkillDiscoveryEntry(uint16 _spellId, float _chance)
+    SkillDiscoveryEntry(uint32 _spellId, float _chance)
         : spellId(_spellId), chance(_chance) {}
 };
